added functions to buttons, can change scenes
