+++ src/main/java/run/StartGame.java
@@ -35,38 +35,6 @@ public class StartGame extends Game {
         this.setScreen(new StartScreen(this));
         state = robot.getState();
 
-
-        /*Input controller*/
-        Gdx.input.setInputProcessor(new InputAdapter(){
-
-            @Override
-            public boolean keyUp(int keycode) {
-                /*input controller*/
-                if(robot.getState() == dead){
-                    System.out.println("You are dead!");
-                    Gdx.app.exit();}
-                else if(robot.getState() == won){ System.out.println("You won!"); System.exit(-1);}
-                else{
-                    switch (keycode){
-                        case Input.Keys.LEFT:
-                            map.moveRobot(robot,Direction.LEFT);
-                            break;
-                        case Input.Keys.RIGHT:
-                            map.moveRobot(robot, Direction.RIGHT);
-                            break;
-                        case Input.Keys.UP:
-                            map.moveRobot(robot, Direction.UP);
-                            break;
-                        case Input.Keys.DOWN:
-                            map.moveRobot(robot, Direction.DOWN);
-                            break;
-                        case Input.Keys.L:
-
-                    }
-                }
-                return false;
-            }
-        } );
     }
 
     private void createState(){
@@ -78,6 +46,7 @@ public class StartGame extends Game {
 
     @Override
     public void dispose(){
+        paint.dispose();
     }
 
     @Override
